requirements.txt
Django==4.2.11
django_environ==0.11.2
djangorestframework==3.15.2
django-cors-headers
drf-registration
django-allauth
dj-rest-auth
Pillow